+++ Nibbled-2-Bits/UI/Panel.h
@@ -7,7 +7,7 @@ class Panel:public UIElement
 private:
 	std::vector<UIElement> m_childlist;
 public:
-	Panel(Play::Point2D pos, float height , float width);
+	Panel(Play::Point2D pos, short height , short width);
 	~Panel();
 
 	void AddToPanel(UIElement& uielement);
